requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
attrs==24.3.0
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
Mesa==2.1.4
Mesa-Viz-Tornado==0.1.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
openai==1.59.8
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.49
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pymdown-extensions==10.14
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
reacton==1.9.0
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rich-click==1.8.5
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
solara==1.43.0
solara-server==1.43.0
solara-ui==1.43.0
stack-data==0.6.3
starlette==0.45.2
text-unidecode==1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
